# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Soko Moto - Dead Hours Flash Deal Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /venues:
    get:
      operationId: listVenues
      tags: [venues]
      summary: List all approved venues
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [restaurant, spa, fitness, bar, experience]
        - name: neighborhood
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, suspended]
      responses:
        "200":
          description: List of venues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Venue"
    post:
      operationId: createVenue
      tags: [venues]
      summary: Create a new venue (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVenueBody"
      responses:
        "201":
          description: Venue created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
  /venues/{id}:
    get:
      operationId: getVenue
      tags: [venues]
      summary: Get a venue by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Venue details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
    patch:
      operationId: updateVenue
      tags: [venues]
      summary: Update venue details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVenueBody"
      responses:
        "200":
          description: Updated venue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Venue"
  /venues/{id}/stats:
    get:
      operationId: getVenueStats
      tags: [venues]
      summary: Analytics for a specific venue
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Venue analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VenueStats"

  # ─── Deals ─────────────────────────────────────────────────────────────────
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users (admin)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # ─── Ratings ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Venue:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        category:
          type: string
          enum: [restaurant, spa, fitness, bar, experience]
        neighborhood:
          type: string
        address:
          type: string
          nullable: true
        imageUrl:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        priceRange:
          type: string
          enum: [budget, mid, upscale, luxury]
        cuisineOrType:
          type: string
          nullable: true
        averageRating:
          type: number
          nullable: true
        totalRatings:
          type: integer
          nullable: true
        status:
          type: string
          enum: [pending, approved, suspended]
        contactPhone:
          type: string
          nullable: true
        contactEmail:
          type: string
          nullable: true
        whatsappNumber:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - category
        - neighborhood
        - status
        - priceRange
        - createdAt
    CreateVenueBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        category:
          type: string
          enum: [restaurant, spa, fitness, bar, experience]
        neighborhood:
          type: string
        address:
          type: string
        imageUrl:
          type: string
        logoUrl:
          type: string
        priceRange:
          type: string
          enum: [budget, mid, upscale, luxury]
        cuisineOrType:
          type: string
        contactPhone:
          type: string
        contactEmail:
          type: string
        whatsappNumber:
          type: string
      required:
        - name
        - category
        - neighborhood
        - priceRange
    UpdateVenueBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: string
        logoUrl:
          type: string
        status:
          type: string
          enum: [pending, approved, suspended]
        cuisineOrType:
          type: string
        contactPhone:
          type: string
    VenueStats:
      type: object
      properties:
        venueId:
          type: integer
        totalDeals:
          type: integer
        totalBookings:
          type: integer
        totalRevenue:
          type: number
        averageFillRate:
          type: number
        averageRating:
          type: number
        topDealType:
          type: string
        revenueRecovered:
          type: number
        bookingsByDayOfWeek:
          type: array
          items:
            type: object
            properties:
              day:
                type: string
              count:
                type: integer
            required: [day, count]
        bookingsByHour:
          type: array
          items:
            type: object
            properties:
              hour:
                type: integer
              count:
                type: integer
            required: [hour, count]
      required:
        - venueId
        - totalDeals
        - totalBookings
        - totalRevenue
        - averageFillRate
        - revenueRecovered

    # ─── Deal ────────────────────────────────────────────────────────────────
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        neighborhood:
          type: string
          nullable: true
        loyaltyTier:
          type: string
          enum: [bronze, silver, gold]
        loyaltyPoints:
          type: integer
        totalBookings:
          type: integer
        totalSpend:
          type: number
        preferredCategories:
          type: array
          items:
            type: string
        isCorporate:
          type: boolean
        corporateCompany:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - phone
        - loyaltyTier
        - loyaltyPoints
        - totalBookings
        - totalSpend
        - isCorporate
        - createdAt
    CreateUserBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
        neighborhood:
          type: string
        preferredCategories:
          type: array
          items:
            type: string
        isCorporate:
          type: boolean
        corporateCompany:
          type: string
      required:
        - name
        - phone
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        neighborhood:
          type: string
        preferredCategories:
          type: array
          items:
            type: string
        corporateCompany:
          type: string

    # ─── Rating ──────────────────────────────────────────────────────────────
...